Set cache item TTL to maximum TTL in the set of its entries.
This will help with out-of-band garbage collection (e.g. redis) of
Guzzle cache entries, as they default to no TTL, even if all of the
entries are expired.

If the set of entries is empty, it sets 1 as the TTL for the cache item,
so that it is expired immediately.